requirements.txt
torch
torchvision
numpy
scipy
streamlit
Pillow
opencv-python
matplotlib
